Merge SymbolPatching library into MemoryModel
A MemoryModel context that is not patched offers very little functionality (throws exceptions for most calls). The SymbolPatcher used to be a larger project including a disassembler but as that is no longer how patching works, there is no value in having a separate project for it.

A MemoryContextInitializer class has been added for simplified context creation with patching.